Added a variant of the SimpleRemote example that uses the Arduino Bounce library.
The Bounce library does the debounce logic, cleaning up the example sketch a lot.
Fortunately its update method still provides the latch facility we need - the
"did the button change state since the last time I checked". We need this so we can
send the whatever-we-say-is-pressed message when it goes from not-pressed to pressed,
and the sendButtonReleased message when it goes from pressed-to-not-pressed.