Add a work-in-progress ethernet sketch for Advanced Remote mode. See caveats below.
Allows a client to connect to port 80 on the Arduino and issue simple commands. Responses from the iPod are returned in JSON format.

Caveats:
The sketch has quite a few commands commented out right now because it appears that I'm running out of RAM, even on a 328.

Sadly it won't even work at all on a 168, even with no commands defined, and the Ethernet shield doesn't work with the Mega unless you hack the SPI stuff.

So on a 328, the result is that a few commands will work fine then it will just hang, and you have to reset.

I added a ClientPrint class to let me send the debug data back to the client but enabling that causes a reset, again possibly due to RAM issues.