Fix byte-shifting on outbound number parameters.
I renamed sendNumber for consistency (we have byte params and we send those with sendByte).

In sendNumber/sendParam I was bit-shifting instead of byte-shifting, and so any value > 255 would not be sent to the iPod correctly. Thus switching to artist 255 would work but switching to artist 256 would return a FEEDBACK_INVALID_PARAM error.