Added support for 'execute-switch' command.
I didn't understand what this was for before, so I didn't implement it. It turns out that this is the command that actually starts playing a selected playlist. It's the equivalent of pressing Play once you've drilled down from Music to a particular track, whether that be via an actual playlist or the artist, genre, etc equivalents. Without this command you could interrogate the iPod but you couldn't actually make it start playing reliably (it would depend on whether it already had an active playlist prior to being put into Advanced Remote mode).

I expanded the ethernet example sketch to include support for this command.